Make navbar reusable in user pages
