+++ frontend/src/pages/Admin.tsx
@@ -1,10 +1,10 @@
 import { Outlet } from "react-router-dom";
-import AdminNavbar from "../components/AdminNavbar";
+import Navigation from "../components/Navigation";
 
 function Admin() {
   return (
     <div>
-      <AdminNavbar />
+      <Navigation />
       <Outlet />
     </div>
